[fix] warnings

+++ undefined/source/include/framebuffer.h
@@ -19,7 +19,7 @@ public:
     /// </summary>
     /// <param name="width">: New width of the framebuffer</param>
     /// <param name="height">: New height of the framebuufer</param>
-    void RescaleFramebuffer(unsigned int width, unsigned int height);
+    void RescaleFramebuffer(float width, float height);
 
 	/// <summary>
 	/// Width of the window
@@ -59,7 +59,7 @@ public:
     /// <param name="height">: Base height of the framebuffer</param>
     /// <returns></returns>
     template <size_t TextureNumber>
-    static Framebuffer* Create(unsigned int width, unsigned int height);
+    static Framebuffer* Create(float width, float height);
 };
 
 #include "framebuffer.inl"
