+++ undefined/source/src/interface/editor_viewport.cpp
@@ -153,7 +153,7 @@ void EditorViewport::RescaleViewport()
 		ViewportCamera->SetPerspective(result);
 	}
 
-	mFramebuffer->RescaleFramebuffer((unsigned int)mWidth, (unsigned int)mHeight);
+	mFramebuffer->RescaleFramebuffer(mWidth, mHeight);
 
 	// TODO add to Renderer
 	glViewport(0, 0, (GLsizei)mWidth, (GLsizei)mHeight);
